FontEditor: Move menu bar into editor and tweak several widgets
Actions are now shared between menu bar and toolbar. Adds an edit
menu to complement toolbar actions. Glyphs are now passed as ints
instead of u8s; fixes Latin Extended+ glyphs failing to update in
real time on map. Converts weight and type to more human-readable
combo box lists. Selected glyph now scrolls into view on load.

+++ Userland/Applications/FontEditor/GlyphEditorWidget.h
@@ -51,7 +51,7 @@ public:
     Gfx::BitmapFont& font() { return *m_font; }
     const Gfx::BitmapFont& font() const { return *m_font; }
 
-    Function<void(u8)> on_glyph_altered;
+    Function<void(int)> on_glyph_altered;
 
 private:
     GlyphEditorWidget() {};
